Extract ContextMenu from sidebar
